Judge ms numbers by mathematical value, as the schema does
Copilot's number-model finding, taken red-first. draft 2020-12's
"integer" is any number whose MATHEMATICAL value is integral — 1000.0
and 1e3 are integer instances a schema-valid fixture may carry — while
json.Unmarshal into int64 judges the spelling, so those fixtures failed
only in the Go driver, against the loader's portable-enforcement
charter. This is the repo's float-spelled-int class again, the one
FlexInt absorbs on the rich-text lane.

Every ms field now parses through one helper: json.Number first as
int64, else as float64 with an integrality check — decidable without
precision loss, since every in-range value sits far below float64's
2^53 exact-integer ceiling (maxScenarioMs ≈ 3.16e11), which the helper
states. A non-integral number is refused in the schema's terms ("is
not an integer: the schema's type is integer — a number whose
mathematical value is integral"), never in encoding/json's; a number
beyond 2^53 is refused as beyond any modeled value. optionalMs (the
five config durations and both envelope members) and advance.ms (a new
scenarioMs type, so the field-level unmarshaler leaves decodeStrict's
unknown-key strictness intact) share the helper.

Red-proven: 1000.0 and 1e3 failed to load against the previous driver,
and 1000.5 was refused with encoding/json's message rather than the
schema's — five cases across the focused spelling test and the
rejection table, all red first, all green after, with the accepted
spellings asserted to decode to exactly 1000.

Gates: both eventfeed packages green, plain and -race; doc-constants
green under LC_ALL=C; gofmt clean.

+++ go/pkg/basecamp/eventfeed/scenario_conformance_test.go
@@ -612,7 +612,7 @@ func (d *driver) nextClientFrame(what string) (clientFrame, error) {
 // AdvanceSettling remains for a caller that genuinely wants a chained firing
 // with an explicit rendezvous. It is deliberately not reachable from a fixture.
 func (d *driver) advance(step *advanceStep) error {
-	if due, ok := d.h.clock.AdvanceIfQuiet(millis(step.Ms)); !ok {
+	if due, ok := d.h.clock.AdvanceIfQuiet(millis(int64(step.Ms))); !ok {
 		return fmt.Errorf(
 			"advance of %dms would fire %v: whether a timer armed by one of those firings lands inside the "+
 				"same window depends on goroutine scheduling, so this script cannot mean the same thing in "+
